refactor: remove app.module.css
