+++ src/components/Logo/Logo.js
