+++ src/components/Logo/index.js
@@ -1,0 +1,3 @@
+import Logo from './Logo';
+
+export default Logo;
