--- src/components/css/app.module.css
@@ -1,28 +1,0 @@
-.wrapper {
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: #fff;
-  background-color: #1a1a1a;
-}
-
-.pageLoading {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  width: 100%;
-  padding: 1rem;
-  font-size: 2rem;
-}
-
-.pageLoading > * + * {
-  margin-top: 1em;
-}
-
-.pageLoading img {
-  width: 100%;
-  max-width: 250px;
-}
